# server/swagger.yaml
openapi: 3.0.3
info:
  title: EventKnit API
  description: |
    EventKnit is a comprehensive ticketing platform API that enables event management, ticket sales, payments, and real-time communications.

    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Base URL
    All API routes are prefixed with `/api/v1`

    ## Rate Limiting
    API requests are rate-limited to prevent abuse. Rate limits vary by endpoint type.
  version: 1.0.0
  contact:
    name: EventKnit Support
    email: support@eventknit.com
  license:
    name: ISC
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.eventknit.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login/registration
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Unauthorized access
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Validation failed
            errors:
              - field: email
                message: Valid email is required
paths:
...
  /auth/register-code/request:
    post:
      tags:
        - Authentication
      summary: Request registration verification code
      description: Request a 6-digit verification code for email-based registration
      operationId: requestRegistrationCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [ATTENDEE, ORGANIZER]
                  default: ATTENDEE
                  description: User role for the new account
      responses:
        '200':
          description: Verification code sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/register-code/verify:
    post:
      tags:
        - Authentication
      summary: Verify registration code and create account
      description: Verify the registration code and complete account creation. Password is required and must contain at least one letter and one number.
      operationId: verifyRegistrationCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: Must contain at least one letter and one number
                firstName:
                  type: string
                lastName:
                  type: string
                phoneNumber:
                  type: string
                role:
                  type: string
                  enum: [ATTENDEE, ORGANIZER]
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /auth/email-oauth/request:
    post:
      tags:
        - Authentication
      summary: Request Email OAuth code
      description: Request a code-based passwordless login/registration
      operationId: requestEmailOAuthCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OAuth code sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/email-oauth/verify:
    post:
      tags:
        - Authentication
      summary: Verify Email OAuth code
      description: Verify OAuth code and authenticate user (creates account if new, logs in if existing)
      operationId: verifyEmailOAuthCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
  /auth/google:
    post:
      tags:
        - Authentication
      summary: Google OAuth login/registration
      description: Authenticate user with a Google OAuth token (ID token or access token from popup OAuth flow)
      operationId: googleAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: Google OAuth token (ID token or access token)
                tokenType:
                  type: string
                  enum: [id_token, access_token]
                  default: id_token
                  description: Type of token being sent
      responses:
        '200':
          description: Google authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthTokens'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Get a new access token using the refresh token stored in an HttpOnly cookie.
        The refresh token is automatically sent via the `refreshToken` cookie set during login.
        The old refresh token is revoked and a new one is issued (token rotation).
      operationId: refreshToken
      parameters:
        - name: refreshToken
          in: cookie
          required: true
          schema:
            type: string
          description: HttpOnly cookie containing the refresh token (set automatically by login/register)
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          expiresIn:
                            type: integer
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/profile:
    put:
      tags:
        - Authentication
      summary: Update user profile
      description: Update the authenticated user's profile information
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                phoneNumber:
                  type: string
                profileImage:
                  type: string
                  format: binary
                bio:
                  type: string
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/password/reset-request:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email to user
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/password/reset-confirm:
    post:
      tags:
        - Authentication
      summary: Confirm password reset
      description: Reset password using reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/password/change:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Change password for authenticated user
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/password/setup:
    post:
      tags:
        - Authentication
      summary: Setup initial password
      description: Set password for users who registered without one (e.g., via OAuth or guest invitation). Only works if the user does not already have a password set.
      operationId: setupPassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: Must contain at least one letter and one number
      responses:
        '200':
          description: Password set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Password already set or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/verify-email:
    get:
      tags:
        - Authentication
      summary: Verify email with token link
      description: Verify user email using a token from an email link
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Email verification token from the link
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email/request:
    post:
      tags:
        - Authentication
      summary: Request email verification code
      description: Send a 6-digit verification code to the user's email for email verification
      operationId: requestEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verification code sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Email already verified or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/resend-invitation:
    post:
      tags:
        - Authentication
      summary: Resend account invitation
      description: Resend the account invitation email for users who haven't set up their password yet
      operationId: resendAccountInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Invitation email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
security:
  - BearerAuth: []